Migrate reviews to store API
